Reorganize rest API routes
Adds a `routes` directory to separate routes based on their concern.
This separates the routes handling agents and batches.
